Initial skaffold & terraform. Add sample workflow
Leverage using skaffold to start up the frontend and backend.
Firestore and Spanner emulator images are included but not yet used

Add a sample ingestion workflow. The sample workflow also uses deploys a
custom "step" which is a Cloud Run service that the workflow calls.

Terraform deploys:
- Firestore database
- Spanner database
- Docker registry
- Sample ingestion pipeline (w/ custom cloud run step)
- Backend
- Frontend

Still need to deploy
- Cloud Task http endpoint
- Public LB for Backend and Frontend
- Alerts

Change-Id: I7cd90f3468fb249f6525b410a5cd5636defb2802