Add new `dev_workflows` target
Background:

Previously, developers could populate their database by running commands
in the DEVELOPMENT.md. It was a lot of lines to copy and paste. And long
term, it was not sustainable as more data sources come on line.

This commit:

Moves the bash script into a Go files:
- util/cmd/local_web_feature_workflow/main.go
  - The executable entrypoint that runs a local version of the workflow
    found at workflows/web-features-repo/workflows.yaml.tftpl
  - This is inspired by https://github.com/web-platform-tests/wpt.fyi/blob/main/util/populate_dev_data.go

Other changes:
- Enable client generation of the openapi yaml files by adding
  openapi/client.cfg.yaml and a new Makefile target that uses the yaml
  file for clients
- Remove the bash script from DEVELOPMENT.md
  - Make a TODO note for a future type of data loading if needed in the
    future.

Change-Id: Ib3e5f461049143554973d792e311482ede699aed