Import wpt.fyi for future wpt data workflow
By importing the wpt.fyi repo, we can easily create client(s) to connect
by using the existing types. However, the import command will be noisy
since it changes the lock files. This commit imports it and then uses it
in one place (in an interface) so the build won't remove it.

Change-Id: Ie59bf9ba4c11784a770fe313a1bde255f081cd81