Create new ci job for playwright
This adds a new playwright test but keeps it separate from the main
precommit. This is because the precommit doesn't need a full setup to work.
By splitting this up, it should allow the first CI job to complete faster.

This also removes setting up the minikube cluster automatically. Now it
is setup when absolutely needed.