Remove unused mock-backend service. (#22)
It is not used.

Previously, I had the thought to use https://stoplight.io/open-source/prism to provide
a mock backend since everything uses OpenAPI. This would allow for a more lightweight
setup for users that only wanted to run the frontend. But getting this started took too
long. So it's best to remove it for now.