Make the sidebar and sidebar menu common
Currently, the naming of the sidebar assumes it is only used on the
overview page. Additionally, the sidebar is rendered on the overview
page explicitly.

This change moves the sidebar up one level to the webstatus-page level.
That way all the CSS can be handled between the sidebar and the rendered
pages can be handled there.

Change-Id: I9120999911e32bd1d99dcc7db08750ecdd195219