Introduce the WPTRun and WPTRunFeatureMetric
`WPTRunFeatureMetric` contains metrics about a web feature that is
calculated during the ingestion pipeline.

ListMetricsForFeatureIDBrowserAndChannel
- Useful for getting metrics over time for a feature. Page: `Feature Detail`

ListMetricsOverTimeWithAggregatedTotals
- Useful for getting metrics over time for a subset of features or all features. Page: `Stats Detail`

`WPTRun` contains common metadata about a run retrieved from the WPT runs
api.

In order to support this, it adds a client that is similar to the
existing datastore client.

Testing:
- Leverage testcontainers library like in the datastore tests to spin
  up a spanner container.

Differences: Spanner does not have a cursor like datastore. So there
will be a cursor

Other notes:
- Update the version of the emulators too
- The initial design used spanner. However, there was a failed attempt to
  use datastore for the same requirements. https://github.com/GoogleChrome/webstatus.dev/pull/28

Change-Id: I9406dbbb790cbbe96ef14c34337a9e93db3fc654