Introduce the spanner client.
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/40
to only include the spanner client. Future PRs will add the different tables.

In the meantime, a placeholder table is there in spanner.sql

client_test uses testcontainers like the datastore client to start the
database and setup the tables.

Also, add the encode and decode cursors which will be used for the WPTRun queries.

Change-Id: I091f221899f11392b9e62bba658ae51a7487e387