Introduce the spanner client. (#43)
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/40
to only include the spanner client. Future PRs will add the different tables.

In the meantime, a placeholder table is there in spanner.sql

client_test uses testcontainers like the datastore client to start the
database and setup the tables.

Also, add the encode and decode cursors which will be used for the WPTRun queries.