Implementations for listAggregatedWPTMetrics & listFeatureWPTMetrics
This commit adds implementations for those two openapi operations.

For now, it's either a 200 for the success or 500 in the event of an error.
Added TODOs for handling other status codes

Also, adjusted the backend binary to allow using the spanner client

Added two helper functions to help deal with optional parameters:
- getPageSizeOrDefault: uses the default 100 in the event the min and max in the openapi document are violated
- getFeatureIDsOrDefault: Handle the default case when no featureIDs (a null object) are passed.

Change-Id: I1c9307779f39b73c6f552787bf6542e08ca40822