Add code to upsert Baseline Statuses
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the code to upsert Baseline Statuses.

The schema for the table is introduced in https://github.com/GoogleChrome/webstatus.dev/pull/59

On updates, the status, low date and the high date can be changed.

The data for this comes from the web features repo.

Change-Id: Ib29a06cf5e9c594ea1bc3c74c2ddd49ac6aff72e