Remove redundant information from Metrics table.
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

Given that the ID from wpt is stored in the runs table, we do not need to
duplicate it in the metrics table. We can leverage the ID that we store instead.

This is a clean up PR that was from the larger PR but we needed to get other tables merged in first.

**Note**
Heads up for reviewers: We will probably need to treat FeatureID as a changeable ID.

That will change the schema after I finish splitting PR 54

https: //github.com/web-platform-dx/web-features/blob/baf32f21b28a33ef8fff8cefc0c43636e5b8ceb6/schemas/defs.schema.json#L10-L23
Change-Id: I4c8231b1b50f33e210fa914b6c1483ae8a473752