Feature: Searching for features in spanner
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the base query to search for pages of features.

This query will be useful for the overview page.

Introduce filterable interface (similar to the filterable for datastore).

Included filters that implement filterable:

- AvailableOn
- NotAvailableOn

If the user provides a list of browsers for either filter,
it will do an OR with that list for that particular filter.

Example. availableOn:browser1,browser2 = availableOn=browser1 OR availableOn = browser2

This behavior may change for the availableOn filter to be an AND.

But the filter query can easily be changed.

Change-Id: I808316fc7b6156d901ffe80b2cd494215bcddd35