Feature: switch to spanner for v1/features endpoint
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54
This is only for the v1/features endpoint not v1/features/{id}. That will come shortly after splitting up PR 54.

Added a new feature search interface to the backend (which spanner implements already). Switched the backend to using that.

Moved the features search handler to a separate file to help with organization.

Added: spanner adapters:
- This creates the translation layer between spanner to the openapi object backend object that is expected.

Change-Id: I1c81be44fff0ab230ded7396ffbcda7e2cddf51c