Add code to insert browser availabilities (#63)
This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the code to insert browser availability for a feature.

The schema for the table is introduced in https://github.com/GoogleChrome/webstatus.dev/pull/59

This data will come from the web features repo. Using the versions there, it will match
up with the versions in the browser releases table to give users when the feature was release.

It will do nothing on update. In the future, we could allow for corrections of specific fields.

Change-Id: I26dcc6bb4fd7f4c83e82d0bcb7cb3b69f2c71786