Use Spanner for v1/features/<id>
This commit uses the same base query for features search
and builds on it to only focus on one feature ID.

Adds the spanner adapter too to translate between spanner and openapi models.

Beforehand, the code used datastore to get feature data. But now we use spanner for all of
the data that is relational.

Change-Id: I01578209f6642c6a2bafddf756d5700a1d424025