Integrate new flexbox classes into all pages (#68)
* Many experiments.

* Add spanner adapters to translate db models to openapi backend models. (#52)

This allows translating of the models from the database to the openapi
models. This code is to adapt the db models to the models in to the
backend models

Other changes:
- Change featureIDs to be an array so that a request like http://<endpoint>/?featureIds=a,b,c can work
- Change the metric count to match int type from the database.
  - Number is a float. Had to change to integer of int64

Change-Id: I42ad13c2bd7c27e8532e1cd3cb01298d0ab15943

* Implementations for listAggregatedWPTMetrics & listFeatureWPTMetrics (#53)

This commit adds implementations for those two openapi operations.

For now, it's either a 200 for the success or 500 in the event of an error.
Added TODOs for handling other status codes

Also, adjusted the backend binary to allow using the spanner client

Added two helper functions to help deal with optional parameters:
- getPageSizeOrDefault: uses the default 100 in the event the min and max in the openapi document are violated
- getFeatureIDsOrDefault: Handle the default case when no featureIDs (a null object) are passed.

Change-Id: I1c9307779f39b73c6f552787bf6542e08ca40822

* Switch from Filter two-level menu to list of buttons (#56)

* Switch from Filter two-level menu to list of buttons

* Add gaps

* Vertical stretch at least to vh.

* Add scrolling of the page-container.

* Replace obsolete hgrow and space-between, and remove from styles.

* Use scrollbar-gutter: stable with overflow: overlay to overlay scrollbar.

* Restructure page to scroll at top level.

* Make the menu toggle the sidebar display.

* Display columns based on ?columns=. (#57)

* Display columns based on ?columns=.

* Addressed review comments

* Adjust the openapi spec for the /v1/features endpoint (#58)

Filters:

Removed:
- baseline_start_year
- baseline_end_year

These filters were examples of how to do queries. We will eventually
bring them back.

Added:
- availableOn
- notAvailableOn

Components:

Added:
- edge to the browsers

Changed:
- score to be a double to match the future value from the database.
- Moved wpt score section to dedicated component called: FeatureWPTSnapshots.
- Added undefined to baseline_status because it can be undefined: https://github.com/web-platform-dx/web-features/blob/1cec26a8adc26203d42ecbef65d4d8a613dab265/schemas/defs.schema.json#L149-L152

Frontend:

Added:
- case for the undefined status to match none.

Change-Id: I04e2d6ba044b0ef1f4ab1b5165be10c37cfadbf2

* Add definition for the rest of the tables (#59)

This adds the rest of the tables for the database.

Also, it introduces the tbls tool to keep documentation up to database
by reading the database to ensure it has the the latest screenshot of data.

Most of this change is the generated documentation.

* Add code to upsert WebFeatures (#60)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the code to upsert WebFeatures.

The schema for the table is introduced in https://github.com/GoogleChrome/webstatus.dev/pull/59

On updates, only the name can be changed.

Change-Id: I6fd23e7a7e01ba0c01a5b2ce372b57358ee2a2dc

* Add code to insert Browser Releases (#61)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the code to insert browser releases.

The schema for the table is introduced in https://github.com/GoogleChrome/webstatus.dev/pull/59

This data will come from https://github.com/mdn/browser-compat-data/tree/main/browsers
Specifically, the github release's data.json.

It will do nothing on update. In the future, we could allow for corrections of specific fields.

Change-Id: Ic876dbf6be8924e540d90c0bd9ef5853f93885a1

* Add code to upsert Baseline Statuses (#62)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the code to upsert Baseline Statuses.

The schema for the table is introduced in https://github.com/GoogleChrome/webstatus.dev/pull/59

On updates, the status, low date and the high date can be changed.

The data for this comes from the web features repo.

Change-Id: Ib29a06cf5e9c594ea1bc3c74c2ddd49ac6aff72e

* Add code to insert browser availabilities (#63)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the code to insert browser availability for a feature.

The schema for the table is introduced in https://github.com/GoogleChrome/webstatus.dev/pull/59

This data will come from the web features repo. Using the versions there, it will match
up with the versions in the browser releases table to give users when the feature was release.

It will do nothing on update. In the future, we could allow for corrections of specific fields.

Change-Id: I26dcc6bb4fd7f4c83e82d0bcb7cb3b69f2c71786

* Remove redundant information from Metrics table. (#64)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

Given that the ID from wpt is stored in the runs table, we do not need to
duplicate it in the metrics table. We can leverage the ID that we store instead.

This is a clean up PR that was from the larger PR but we needed to get other tables merged in first.

**Note**
Heads up for reviewers: We will probably need to treat FeatureID as a changeable ID.

That will change the schema after I finish splitting PR 54

https: //github.com/web-platform-dx/web-features/blob/baf32f21b28a33ef8fff8cefc0c43636e5b8ceb6/schemas/defs.schema.json#L10-L23
Change-Id: I4c8231b1b50f33e210fa914b6c1483ae8a473752

* Feature: Searching for features in spanner (#65)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54

This introduces the base query to search for pages of features.

This query will be useful for the overview page.

Introduce filterable interface (similar to the filterable for datastore).

Included filters that implement filterable:

- AvailableOn
- NotAvailableOn

If the user provides a list of browsers for either filter,
it will do an OR with that list for that particular filter.

Example. availableOn:browser1,browser2 = availableOn=browser1 OR availableOn = browser2

This behavior may change for the availableOn filter to be an AND.

But the filter query can easily be changed.

Change-Id: I808316fc7b6156d901ffe80b2cd494215bcddd35

* Feature: switch to spanner for v1/features endpoint (#66)

This is splitting https://github.com/GoogleChrome/webstatus.dev/pull/54
This is only for the v1/features endpoint not v1/features/{id}. That will come shortly after splitting up PR 54.

Added a new feature search interface to the backend (which spanner implements already). Switched the backend to using that.

Moved the features search handler to a separate file to help with organization.

Added: spanner adapters:
- This creates the translation layer between spanner to the openapi object backend object that is expected.

Change-Id: I1c81be44fff0ab230ded7396ffbcda7e2cddf51c

* Script: Generate fake data. (#67)

Adds a makefile target to generate fake data.

This should be used in a separate terminal after the local environment is up with `make start-local`.

It will temporarily open the port into the local cluster, run the script to load data, and close the port.

Change-Id: I5d4faec9268077c07a2f109beb8f11c76dc3f89b

* Cleanup and hide vertical scroll of sidebar

* More cleanup

* Fix playwright tests and update snapshots

---------

Co-authored-by: James C Scott III <jcscottiii@users.noreply.github.com>
Co-authored-by: Jason Robbins <jrobbins@gmail.com>