Make necessary changes to push to GCP (#78)
It has been awhile since the last push to GCP. Now that we use spanner, I found
out some differences between spanner emulator and real spanner. Additionally, I fixed
issue regarding the website talking to the api behind UberProxy.

More below details below:

The fontend client was fixed to prevent headers from being automatically
set by the openapi generated client. This is needed for UberProxy.

Modified DEVELOPMENT.md with more steps on how to deploy to staging.

Modified the fake data script to be configurable. This allows us to load
data fake data into staging.

Modified terraform to give backend access to spanner with IAM membership.

Modified table to add more NOT NULL constraints. (Was an attempt to fix issue 77 but might as well keep it.)

Modified features search base query to wrap the metrics in TO_JSON. That is an issue that only occurred on real spanner not in the emulator.
Issue #77 was created to handle it.

Change-Id: Idfc0680b4808456b10df35d4cb417ec272edc966