Update Web features consumer pipeline to use spanner tables
Previously, the workflow to consume web features involved two microservices
1. A Generic Repo downloader which downloaded the web features repo and sent each individual file to
2. The web feature consumer workflow (which used datastore)

Now that the data can come from the github release, we do not need step 1 anymore (there will be more clean up after this)

This change implements a common library to get assets from GitHub Releases (which will be useful the BCD data too)

This change implements logic to download the asset, parse it and then store it.

Modified the openapi yaml for the web consumer api to be simple. Modified the local workflow runner to use the simple generated openapi models.

Fixes:
- The prefilter was building a filter when the database is empty and would generate invalid SQL. That has been fixed.

Other things:
- Added some initial error logging to help out in the future (was useful for the fix above)
- I will also clean up the datastore usage in the near future. (unless we think we may need it)

Change-Id: Ie1450eabbc2d8aab3960592c451c2873dec52e04