Add Spanner to web feature ingestion. Fix prefilter (#113)
Previously, the web features ingestion pipeline only connected to datastore.

This change adds the details so that the ingestion pipeline can connect to spanner, locally and in GCP.

This will prepare us for loading the web features data.

Additionally, the prefilter logic for GCP spanner has a bug where it prints out a string if the database is empty.
That should not happen. Fixed that. And added a test to catch that in the future.

Other touch ups:
- Fix command in DEVELOPMENT.md. And remove comment since you don't need to save the ENV_ID since it is "staging"