Use the new helpers for the web feature workflow
This change uses the changes from #115 in the entrypoint of the web consumer workflow.

Now the workflow only has three stages inside the same binary and no longer relies on the input from another step.

Those steps are:
- Get the latest data.json
- Parse the data
- Store the data

The workrflow template file has been updated to show that it only needs to call one process now.

The openapi.yaml was modified to remove the need for any input from google storage bucket too. The dev_workflow
command is edited to remove references to the repo downloader binary. And it also edited the generated models to fit
the openapi yaml.

Additionally, since we do not need to use google storage anymore, I started to remove it
from the web_feature_consumer service. A bigger clean up will happen after this to remove from
the overall infrastructure.

Change-Id: Iaf639ff5d1f260c87ddc3aaf764f95aa4c46b10f