Fix playwright in CI & load data upon playwright run
This hopefully fixes the Github CI issues of where the action runs out of space.

Additionally, I removed the pixel diff. Since we are using the devcontainer now,
there will be no diffs between CI and local. (whereas previously the CI did not
use devcontainer for the playwright action).

This also loads up the data so that we can get some better screenshots.

Change-Id: I6449481282844bfd8a7fee12c486c952d3f8afcf