Add local In-Memory Cache helper
This implements the future Cache interface that will be used by the WPT Workflow.

The WPT workflow will need to cache a version of the data.json from the web features Github releases
to prevent each local worker from downloading and potentially getting throttled.

It uses generics so that it can used for any data types in the future.

The local in memory cache is thread safe.

Change-Id: Ib67f217d13f1718dc43925af9e245833774094e9