Add HTTP Results Getter for WPT Run results.
This is the implementation of the ResultsGetter that actually makes the call
to the HTTP endpoint for the results from the GCP bucket.

After getting the results, it will only try to parse the latest v2 format.

The v2 format was introduced around Sept 2022 [1]. So that should get us a good amount of data.

That is good to know for the workflow to not consume data before then.

[1] https://github.com/web-platform-tests/wpt.fyi/commit/cee243faf76719f174c8b74f6bdeb5ed59d14280

Change-Id: I9aef40e55857219928133d63891504a1c5e9d111