Add lightweight worker pool
The WPT workflow runs locally. During the split up of #118, I realized
that the entrypoint logic essentially was a workerpool manager and it could
be used for other future workflows if we specify generic arguments for the inputs.

That is what this does and puts it in the shared `lib` directory to be used by
the WPT workflow and future workflows.

Change-Id: I867338f66c0f101ae64fe719138ef14b30dc3e0c