Add job binary for WPT workflow
This is the entrypoint to the WPT workflow that builds on all the work of
the others PRs split out from PR #118

It parses all the needed configurations, creates the worker pool,
sends the jobs and reports back any errors.

Additionally, the Makefile has been changed so that the dev_workflow also
calls this new job.

The job.yaml file describes the kubernetes job.

Other changes: changes to the GetWebFeaturesData method to match the
return type of the datacacher type.

Change-Id: I1f31f64168048104f513427fe0989c1c75b850e7