Add infrastructure to deploy the WPT job
This change adds the terraform to deploy the job.

It leverages some things from the existing cloud run services.

However, this is the first cloud run job. In the future, we can look at
how to move some of these things to common terraform modules.

Also added the template for the workflow. It is a simple workflow
that just calls the cloud run job.

Other changes:
- Reformat the terraform vars file for staging
- Turn on the cloud scheduler api for the future when we want to do cron triggers of the workflow

Change-Id: I681d9d8266a4a8ebf98b545e093d6cdf3f1a97ab