Add offset based pagination and total counts
The UI needs to show both
1) the total number of features and
2) allow users to visit the website by specifying an offset (along with the existing page size)

For requirement 1, this change adds a new query that gets the full count of
elements with the same queries but without the limits

For requirement 2, this change enhances the page token to allow offset based pagination.
Offset based pagination in general is not efficient but it makes the most sense for the UI
However, after every query for features, we still return the existing cursor based pagination instead.
The end user is responsible for creating its own offset token.

Change-Id: If505127e18cd435d5273fefce2f10d14a7792afd