Add PageMetadataWithTotal component in openapi
The previous PageMetadata only had the next page token.

Instead of adjusting the current page metadata to have an optional total,
it would mean we would need to document which endpoints will return total
versus not.

That would be tedious to maintain as we add more endpoints.

Instead, this change adds a new component that has both the next page token
and total which lets the client know that this client is always returning the total.

Change-Id: I78317b46fbb427a63d07cd4a3995a8972f16edf2