Add logic to handle DESC in addition to ASC order
When the sortable is created, we also store whether it is asecending or descending

With that information, we can store information about the operator to use for the next page

Upon parsing a token, we check if it is indeed an expected operator. And then fill in the
appropriate operator.

Change-Id: Id6aafcc9a040243c8390a3075234de7427ff4cd1