Add JSONSchema for BCD Browser Releases
This adds the json schema to the repository for the BCD Browser Release data.

The data.json is the union of these two schemas:
- https://github.com/mdn/browser-compat-data/blob/main/schemas/browsers.schema.json
- https://github.com/mdn/browser-compat-data/blob/main/schemas/compat-data.schema.json

We only need the first one for now to fill in the Browser Releases table.

Besides adding it, added commands to download and generate the Go code for it.

This process is similar to the web features json schema.

Change-Id: Ib23f73838d5edc8d8bc94f51d1d326938333af81