Offset pagination token for v1/features calls in frontend
This change adds logic to create the offset pagination tokens in the frontend

Given that it is not the default page token for the endpoint, I added
a disclaimer for outside readers who may decide to do the same thing.
That is we do not guarantee that the offset token structure will stay the same.
Instead API users should rely on the default page token that is returned.

Change-Id: Ia99f6d287a921b2d941e41c094a458e3cdca4de6