Refactor the tests for feature search & get feature
The tests previously listed out the features multiple times.
This caused the test file to be really repetitive and length.
And simple modifications caused multiple lines to be updated in
the test file.

This combined with the fact that we will be storing different
WPT test metric types soon means we should refactor the tests now.

I refactored the tests to have a simple getter method
to retrieve the feature needed. This getter will allow the caller
to specify which metric type is needed.

Also, I moved some of the tests to a table tests structure to help DRY up
some of the test methods.

No functionality is actually modified in this change.

Change-Id: I0c3400d203574a33f98750af5aad2654360a1c94