Refactor feature search query to use text/template
Before making further changes to the feature search
query, it would be great to use a more powerful string builder
such as the text/template.

We continue to use params for any values. But for the column names
that cannot use spanner params, we leverage the text/template
templating language to pass those in.

The templates are compiled once at module start up time in the init() method

Added a simple helper method called metricsPassRateColumn that returns
the column name for the type of metrics we want. We will soon support
different types of metrics and this method can be modified to accomodate that.

Change-Id: I28bd29029e312fbfc629d739acf08288b47f727c