Add support for subtest wpt metric view
Previously, the dashboard only stored the Test View data in the Metrics table

This change modifies the table schema to also store the Subtest data from wpt.fyi
- In the future, we could also store the fractional interop view too

This data is calculated in an updated score_webfeature.go file. This logic is taken from
this section:
- https://github.com/web-platform-tests/wpt.fyi/blob/ea1f3f64aafde5ff5a4c4102ab150cdefa25e917/webapp/views/wpt-results.js#L719-L795

Also, updated the existing scoreTest to more closely match what is happening by checking
explicitly for the status before incrementing.

The spanner database query for FeaturesSearch/GetFeature is updated to use
metricsPassRateColumn and metricsPassRateIndex which depend on the desired metric.

The openapi document has been updated so that /v1/features and /v1/features/{id}
can optionally pass in the desired view of the data. If none is provided or an invalid
value is provided, we use subtest view.

Notables:
- The list metric endpoints for charts still show the test view data and not subtest.
  - A future change will address that.

Change-Id: Ife6e33e91eb3db9b42ba335510bc7fc178d02b18