Get implementation status from database
This change implements the changes needed to get the implementations status from
the database and passes it through the layers back to the handler

The database query does a similar query to the metrics where it creates an
array of structs. We do the similar thing where we put the placeholder item in
there to prevent the error where it detects an array with a null value inside of it.

After getting it from the database, the adapter converts it for the backend to serve back.

Change-Id: Id67b72bb95a9d627fea7b9dbbc695e65bf3bad1d