Use internal FeatureID for database Foreign Keys
Some of the tables relied on the the feature id from web-features repo
as their key. However, it turns out that ID is not stable and subject to
change.

Luckily, we already had an internal UUID ID in the web features table just
in case.

This change moves the foreign key from the external feature ID to the UUID

This requires an extra join for some queries. Additionally, it requires a lookup
query to translate the external ID into the internal ID before insert/updates.

We already do this pattern with WPT data where we have the ID from wpt.fyi but rely
on our own ID for the primary key.

Change-Id: Ic39f0b45b8f891b314928225c2aa01bbbde42aab