Support baseline_date search term (#184)
This PR adds support for `baseline_date`. Currently, this only supports
the discussed range operator `..`. The range is inclusive on both the start
and end date to mimic what GitHub does for search as discussed too.

This term needed support for the `>=` and `<=` on the backend. So these operators were added.

Other changes:

During the implementation of this, I saw that I needed to have split up AND & OR from the new
operators. AND & OR (and ROOT and NONE) are now called keywords.

As a result of adding the `>=`, `<=` operators, I also added `=`, `!=`, `>`, and `<`.

These operators impact the term and a field called Operator was introduced to the SearchTerm object.

The `=` is the default operator that was used for the existing terms so existing code was

It became hard to handle negation with these new operators with the existing code. The visitor code
calls Invert() on the Keyword and Operators to negate them. The feature search tests proved that this yielded
the same expected results.