Make subtest_count the default for timeseries data
The overview page was already switched over to use the subtest count.
However, the time series data still had test count as the metric.

This change modifies the timeseries queries to allow the metric columns
to be configurable like the overview page query.

The openapi document was modified so that the metric view can switch
between test_counts and subtest_counts. Previously, it only allowed subtest_counts

In the future, we can easily switch over to test count now when ready.

The frontend now explicitly calls for subtest count too for the metrics over time.

More changes:
- Previously, the querying logic for the metrics over time was contained
  in many strings that were formed with fmt.Sprintf. Now all those queries
  have been refactored into a single string template.

Other considerations:
- We can probably get rid of `/v1/features/{feature_id}/stats/wpt/browsers/{browser}/channels/{channel}/{metric_view}`
  and rely on `/v1/stats/wpt/browsers/{browser}/channels/{channel}/{metric_view}` and pass in a single
  feature.

Change-Id: Id46f97560bf9061440dadeb92f74950f4db85bfa