Implement UI for baseline date bookmark query (#187)
* Implement baseline date bookmark query

Now that the backend supports baseline date bookmark queries,
we can have the bookmarks fill in the url which will automatically
fill in the search bar and search.

Added getLocation and naviage methods to the sidebar menu
that can be changed by the tests.

This is because sinon cannot stub es modules. The import map suggestion
in the docs [1], did not work either. As a result, we can use this form of dependency
injection to set the stub.

Added helper methods navigateToUrl and getCurrentLocation to the router file.

This file can be used to
1. centralize all usage of `window` in there.
2. Allow for mocking.

[1] https://modern-web.dev/docs/test-runner/writing-tests/mocking/#mocking-es-modules

Change-Id: I2f6b0cba0746477c59e287f449de65bf6deef512

* address feedback

Change-Id: Id4e3898d2231a82bfef263f73ab516203d6a30cc

* remove handle click

Change-Id: Ie6fb136a66c1c9c4044920b7fcfca60ac7b88b3f

* cleanup

Change-Id: Iaab0409d5dd75f78ef838af6363cf98d5f5ddcf1