Set CI to trigger on merge queue (#198)
Currently, we require devs to have the latest branch before merging

This ensures that things like screenshots always have
the correct golden files. But a lot of changes don't affect
the screenshots so it slows down developer time because they
have to continually update their branch before it merges.

We can leverage [merge queues](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue)
to prevent that. It will create a temporary branch with the latest changes
for the devs upon merging their PR. It will ony fail if:
- It can't pull in the latest changes to your branch
- The CI fails with the latest change and your PR

So no more updating your branch manually. It will be handled automatically.

I will turn on merge queues after this change is merged.

Change-Id: I8f2d720340c1c2dfef9296336bb16b64edfa8649