Modify query to get the latest metric for each feature
Previously, the query got the last run for the browser then got
all the metrics for that run. But sometimes runs are incomplete runs

As a result, some metrics were missing

Instead, we now get the latest metric for that browser, skipping any
incomplete runs.

Fixes #205

Check that issue for more details

Change-Id: I9dcd834cdbeb8ee63371a716b87d2087c3e21be4