Ignore subtests for a feature if there is a crash
In WPT, if a test crashes, it does not record the correct number of subtests (it does for tests though).

Instead of recording the wrong number for subtests from WPT, ignore that feature's subtests count

For the timeseries data that is returned, ignore those data points.
For the feature search which gets the latest, it will get the latest data (so it could be null)
- This prevents the search from having to go back N runs for a given feature to find a non null metric.

Change-Id: I64a74ba12cb979185a2623d7360fd9b7cda58f4b