Consume the spec information from web features (#215)
There are certain features that will never have a WPT score. And the UI
will need a way to know which ones those are. Currently, the only information
we have available is the spec links have a prefix to indicate that those
features indeed will not have a WPT score.

This change modifies the pipeline so that we can store and read that information

Future changes include:
- Change to the feature search / get feature queries to return that information.

Change-Id: I2015083eb32180df16259369b17bc50bcb2889eb