Modify terraform to deploy redis. (#227)
This change modifies terraform to deploy redis.
Given redis is a regional service and we are multi-region, we leverage the existing
region-to-subnet map to know which regions to create redis in.

In the meantime, we can have a quick TTL for redis.

An issue to follow up on this in #225 has been filed.

We can probably increase the TTL for prod soon. But while data is still changing before launch, we
can keep it quick. Open to suggestions for the initial TTL.

About the terraform changes itself:
- Because the project spans multiple projects, we have a shared VPC. This is because
  Memorystore is different from other services like spanner. Spanner instances do
  not live in our IP space. But Memorystore instances do. As a result, we have to set up
  "Private Services Access" by giving it a slice of IPs for the service to deploy the service
  into. More here in the blue banner at the top of [1] and [2].
- The TTL for prod and staging are set to 5minutes. We can increase this in the future as we
  get closer to launch.

Other changes:
- Tell terraform to enable needed APIs.

[1] https://cloud.google.com/memorystore/docs/redis/networking
[2] https://cloud.google.com/memorystore/docs/redis/networking#private_services_access

Change-Id: I47e8d7f5c8eee7f0e0ab90a2f8fe6c88bc79933e