Ingest and expose non-relational web feature metadata
There exists additional metadata that we always wanted to ingest for a feature.
Such as a the description, html-description, CanIUse IDs.

However, putting that information in spanner doesn't make too much sense
given that we do not need to do any relational joins against it. Plus,
the data is likely to change shape. This makes maintenance of this data very
hard.

As a result, we put this in something more suitable, datastore.

This change begins the ingestion of the the can i use IDs and description.
We also want to store the html description but we should discuss that to make
sure render that safely.

On the API side, a new endpoint exists /v1/features/{feature_id}/feature-metadata to expose that
data. For now, it only returns Can I Use IDs. The returned object, while bulky, is structured in case we want to add any information
without breaking our schema in the future.

Other changes:
- Update UpsertWebFeature to return the generated ID and use that as the
  web feature id in datastore.
- Update the defs.schema.json file from the web features repo. This allows us to get the description.

Change-Id: I0783787531c9e183c8c8d9b7759f1285cceac3d5