Expose feature metadata endpoint
On the API side, a new endpoint exists /v1/features/{feature_id}/feature-metadata to expose that data.
For now, it only returns Can I Use IDs. The returned object, while bulky,
is structured in case we want to add any information without breaking our schema in the future.

This change also introduces the adapters which convert the database model from datastore
to the expected backend model that is returned

Future changes:
- Add more tests to backend/pkg/httpserver/get_feature_metadata_test.go

Change-Id: Ia044d282c97054131c83706c28026c6d27a677c0