Fixes various nginx issues
Fixes #235 and should not break the existing 404 e2e tests

Additionally, add the errors from openapi into the client.
Then use it for the feature detail page.

Previously, the feature detail page shows this message for any error (including 404):
- Error when loading feature <featureKey>.

But given we have a not found page, a first step would be to leverage that by
forcing the browser to navigate to a bad 404 route with the `errors-404` path prefix.

In the future, we can add separate path prefixes in the nginx config for different status codes, for example
/errors-500/ would make the page render with a 500 status code.

Change-Id: I9ef1950cef956f8a3005b9329559f654e5b3a9b9