Store misc. info for feature in a wpt run and expose
We want to have more insights into what is going on for the latest run.

This change adds a new column called FeatureRunDetails. It is a JSON blob
of information that can be stored. This is meant to be temporary until
we figure out other types of information that we want to show. So instead
of created a new column for each new piece of information, it can be stored
in this JSON column.

The API layer exposes a new field called Metadata via setting additionalProperties
to true. (allows any mapping of string to any value type). Similarly, this
allows to not commit to keeping this field long term. A note has been added
to the description. If a value continues to stay, we can promote it to its
own value.

When is this used:
- In the WPT consumer workflow, if there is a crash, it stores a mapping "status" -> "C"

What's next:
- Have the frontend check for this and render appropriately.

Change-Id: I0a0c9a08481ed65ecdaa22de81d3ba330e082369